Revert "Merge tag 'LA.UM.8.3.r1-08500-sdm845.0' into new-caf"
This reverts commit c628e228af60599c2eb2e70a9f36e227796ddfb0, reversing
changes made to 97c13ce5454935032be9bc1b625128b6bb6e2814.

+++ fs/file.c
@@ -89,7 +89,7 @@ static void copy_fd_bitmaps(struct fdtable *nfdt, struct fdtable *ofdt,
  */
 static void copy_fdtable(struct fdtable *nfdt, struct fdtable *ofdt)
 {
-	size_t cpy, set;
+	unsigned int cpy, set;
 
 	BUG_ON(nfdt->max_fds < ofdt->max_fds);
